Decode to UTF-8 before logging
This fixes an exception where a header contains a character that cannot
be converted to ASCII, e.g. 0xea.  Python will attempt to encode unicode
strings to ascii before passing them to b64encode (which passes it to the
internal C function binascii.b2a_b64).  This commit forces encoding of
the string as UTF-8, which contains no characters that cannot be encoded.